+++ crates/core/utils/src/fs.rs
@@ -1,33 +1,11 @@
 use async_recursion::async_recursion;
 use futures::future::try_join_all;
-use json_comments::StripComments;
 use moon_error::{map_io_to_fs_error, map_json_to_error, map_yaml_to_error, MoonError};
-use regex::Regex;
 use serde::de::DeserializeOwned;
 use serde::Serialize;
-use std::io::Read;
 use std::path::{Path, PathBuf};
 use std::time::{Duration, SystemTime};
 use tokio::fs;
-
-#[inline]
-pub fn clean_json<T: AsRef<str>>(json: T) -> Result<String, MoonError> {
-    let json = json.as_ref();
-
-    // Remove comments
-    let mut stripped = String::with_capacity(json.len());
-
-    StripComments::new(json.as_bytes())
-        .read_to_string(&mut stripped)
-        .map_err(MoonError::Unknown)?;
-
-    // Remove trailing commas
-    let stripped = Regex::new(r",(?P<valid>\s*})")
-        .unwrap()
-        .replace_all(&stripped, "$valid");
-
-    Ok(String::from(stripped))
-}
 
 #[inline]
 pub async fn copy_file<S: AsRef<Path>, D: AsRef<Path>>(from: S, to: D) -> Result<(), MoonError> {
@@ -185,7 +163,7 @@ where
 
 #[inline]
 pub async fn read_json_string<T: AsRef<Path>>(path: T) -> Result<String, MoonError> {
-    clean_json(read(path).await?)
+    crate::json::clean(read(path).await?)
 }
 
 #[inline]
@@ -407,26 +385,6 @@ pub mod sync {
         let data = read_to_string(path).map_err(|e| map_io_to_fs_error(e, path.to_path_buf()))?;
 
         Ok(data)
-    }
-
-    #[inline]
-    pub fn read_json<P, D>(path: P) -> Result<D, MoonError>
-    where
-        P: AsRef<Path>,
-        D: DeserializeOwned,
-    {
-        let path = path.as_ref();
-        let contents = read_json_string(path)?;
-
-        let json: D = serde_json::from_str(&contents)
-            .map_err(|e| map_json_to_error(e, path.to_path_buf()))?;
-
-        Ok(json)
-    }
-
-    #[inline]
-    pub fn read_json_string<T: AsRef<Path>>(path: T) -> Result<String, MoonError> {
-        clean_json(read(path.as_ref())?)
     }
 
     #[inline]
